Static -> Memoria
Cambio de la clase llamada Static a Memoria
