+++ src/UI/Cliente/MPreferenciasCliente.java
@@ -1,6 +1,6 @@
 package UI.Cliente;
 
-import Logica.Static;
+import Logica.Memoria;
 
 public class MPreferenciasCliente extends javax.swing.JFrame {
 
@@ -11,13 +11,13 @@ public class MPreferenciasCliente extends javax.swing.JFrame {
 
     private void settings() {
         //Establece el icono en la barra de estado y en el icono.
-        setIconImage(Static.getIconImage());
+        setIconImage(Memoria.getIconImage());
         //Coloca el frame en el centro de la pantalla.
         this.setLocationRelativeTo(null);
         //Hace el frame visible.
         this.setVisible(true);
         //Establece texto en la barra de estado.
-        this.setTitle(Static.app_name);
+        this.setTitle(Memoria.app_name);
         //No dejar que el frame se pueda hacer de tamaño grande
         this.setResizable(false);
     }
